+++ lib/main.dart
@@ -4,6 +4,7 @@ import 'package:flutter/services.dart';
 import 'package:lifeline/screens/blood_donation_screen.dart';
 import 'package:lifeline/screens/health_record_screen.dart';
 import 'package:lifeline/screens/medical_history_screen.dart';
+import 'package:lifeline/screens/terms_and_conditions_screen.dart';
 import 'package:lifeline/screens/user_login_screen.dart';
 import 'package:lifeline/screens/user_profile_screen.dart';
 import 'package:lifeline/screens/user_registration_screen.dart';
@@ -57,6 +58,7 @@ class LifeLine extends StatelessWidget {
         UserSearchScreen.id: (context) => UserSearchScreen(),
         HealthRecordScreen.id: (context) => HealthRecordScreen(),
         MedicalHistoryScreen.id: (context) => MedicalHistoryScreen(),
+        TermsAndConditionsScreen.id: (context) => TermsAndConditionsScreen(),
       },
     );
   }
