+++ lib/screens/user_registration_screen.dart
@@ -1,7 +1,10 @@
+import 'package:auto_size_text/auto_size_text.dart';
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:lifeline/components/custom_text_field.dart';
 import 'package:lifeline/components/rounded_button.dart';
 import 'package:lifeline/constants.dart';
+import 'package:lifeline/screens/terms_and_conditions_screen.dart';
 import 'package:lifeline/screens/user_dashboard_screen.dart';
 import 'package:lifeline/services/authenticate.dart';
 import 'package:modal_progress_hud/modal_progress_hud.dart';
@@ -19,6 +22,7 @@ class _UserRegistrationScreenState extends State<UserRegistrationScreen> {
   TextEditingController password = new TextEditingController();
 
   bool loadingIndicator = false;
+  bool checkedValue = false;
 
   @override
   Widget build(BuildContext context) {
@@ -60,30 +64,67 @@ class _UserRegistrationScreenState extends State<UserRegistrationScreen> {
                     obscureText: true,
                     controller: this.password,
                   ),
+                  CheckboxListTile(
+                    title: RichText(
+                      text: TextSpan(children: [
+                        TextSpan(
+                          text: 'I agree to the ',
+                          style: kTextStyle.copyWith(
+                            fontSize: 16.0,
+                            fontWeight: FontWeight.w400,
+                          ),
+                        ),
+                        TextSpan(
+                          text: 'terms and conditions',
+                          recognizer: TapGestureRecognizer()
+                            ..onTap = () {
+                              Navigator.pushNamed(
+                                  context, TermsAndConditionsScreen.id);
+                            },
+                          style: kTextStyle.copyWith(
+                            color: Colors.green,
+                            decoration: TextDecoration.underline,
+                            fontSize: 16,
+                            fontWeight: FontWeight.bold,
+                          ),
+                        )
+                      ]),
+                    ),
+                    value: checkedValue,
+                    onChanged: (newValue) {
+                      setState(() {
+                        checkedValue = newValue;
+                      });
+                    },
+                    controlAffinity: ListTileControlAffinity
+                        .leading, //  <-- leading Checkbox
+                  ),
                   RoundedButton(
                     text: 'Register',
                     color: Colors.green[900],
                     onPressed: () async {
-                      setState(() {
-                        loadingIndicator = true;
-                      });
-
-                      try {
-                        final user = await Auth()
-                            .register(this.email.text, this.password.text);
-                        if (user != null)
-                          Navigator.pushNamed(context, UserDashboardScreen.id);
+                      if (checkedValue) {
                         setState(() {
-                          loadingIndicator = false;
+                          loadingIndicator = true;
                         });
-                      } catch (e) {
-                        print(e);
-                        Toast.show(
-                          e.message,
-                          context,
-                          duration: Toast.LENGTH_LONG,
-                          gravity: Toast.TOP,
-                        );
+                        try {
+                          final user = await Auth()
+                              .register(this.email.text, this.password.text);
+                          if (user != null)
+                            Navigator.pushNamed(
+                                context, UserDashboardScreen.id);
+                          setState(() {
+                            loadingIndicator = false;
+                          });
+                        } catch (e) {
+                          print(e);
+                          Toast.show(
+                            e.message,
+                            context,
+                            duration: Toast.LENGTH_LONG,
+                            gravity: Toast.TOP,
+                          );
+                        }
                       }
                     },
                   ),
